+++ src/image/MagickImageSink.cpp
@@ -16,7 +16,7 @@ void MagickImageSink::append(const std::vector<uint32_t> &pixels) {
 }
 
 void MagickImageSink::write() {
-    Magick::Image image(width, imagePixels.size()/width, "BGRA", Magick::CharPixel, imagePixels.data());
+    Magick::Image image(width, static_cast<unsigned int>(imagePixels.size()/width), "BGRA", Magick::CharPixel, imagePixels.data());
     image.quality(100);
     image.opacity(0);
